+++ addon/components/flexberry-lookup.js
@@ -532,6 +532,7 @@ export default FlexberryBaseComponent.extend({
     'title',
     'lookupLimitPredicate',
     'relatedModel',
+    'updateLookupAction',
     '_modalDialogSettings',
     '_lookupWindowCustomPropertiesData',
     function() {
@@ -548,6 +549,7 @@ export default FlexberryBaseComponent.extend({
         perPage: perPage || this.get('perPage'),
         folvComponentName: this.get('folvComponentName'),
         modalDialogSettings: this.get('_modalDialogSettings'),
+        updateLookupAction: this.get('updateLookupAction'),
       };
     }),
 
@@ -918,6 +920,16 @@ export default FlexberryBaseComponent.extend({
   },
 
   /**
+    Triggered by the component template if the component has a block form definition.
+
+    @private
+    @property _hasBlockSetter
+  */
+  _hasBlockSetter: Ember.computed(function() {
+    this.set('_hasBlock', true);
+  }),
+
+  /**
     Called after a component has been rendered, both on initial render and in subsequent rerenders.
     [More info](http://emberjs.com/api/classes/Ember.Component.html#event_didRender).
 
@@ -930,6 +942,16 @@ export default FlexberryBaseComponent.extend({
     let isDropdown = this.get('dropdown');
     if (isAutocomplete && isDropdown) {
       throw new Error('Component flexberry-lookup should not have both flags \'autocomplete\' and \'dropdown\' enabled.');
+    }
+
+    let hasBlock = this.get('_hasBlock');
+
+    if (isDropdown && hasBlock) {
+      throw new Error('Component flexberry-lookup should not have both flag \'dropdown\' enabled and the block form definition.');
+    }
+
+    if (isAutocomplete && hasBlock) {
+      throw new Error('Component flexberry-lookup should not have both flag \'autocomplete\' enabled and the block form definition.');
     }
 
     let cachedDropdownValue = this.get('_cachedDropdownValue');
@@ -1163,7 +1185,8 @@ export default FlexberryBaseComponent.extend({
               {
                 relationName: relationName,
                 modelToLookup: relatedModel,
-                newRelationValue: result.instance
+                newRelationValue: result.instance,
+                componentName: _this.get('componentName')
               });
           });
         } else {
@@ -1225,7 +1248,8 @@ export default FlexberryBaseComponent.extend({
                       {
                         relationName: relationName,
                         modelToLookup: relatedModel,
-                        newRelationValue: record
+                        newRelationValue: record,
+                        componentName: _this.get('componentName')
                       });
                   }
                 });
@@ -1349,7 +1373,8 @@ export default FlexberryBaseComponent.extend({
           {
             relationName: relationName,
             modelToLookup: relatedModel,
-            newRelationValue: newValue
+            newRelationValue: newValue,
+            componentName: _this.get('componentName')
           });
       }
     });
@@ -1546,7 +1571,8 @@ export default FlexberryBaseComponent.extend({
           {
             relationName: relationName,
             modelToLookup: relatedModel,
-            newRelationValue: record
+            newRelationValue: record,
+            componentName: _this.get('componentName')
           });
       }
     });