+++ addon/components/flexberry-objectlistview.js
@@ -5,6 +5,7 @@
 import Ember from 'ember';
 import FlexberryBaseComponent from './flexberry-base-component';
 import { translationMacro as t } from 'ember-i18n';
+import runAfter from '../utils/run-after';
 
 /**
   Component to view list of object.
@@ -177,6 +178,15 @@ export default FlexberryBaseComponent.extend({
     @type Number
   */
   hierarchicalIndent: undefined,
+
+  /**
+    Using `ember-test-selectors`, creates `[data-test-component=flexberry-objectlistview]` selector for this component.
+
+    @property data-test-component
+    @type String
+    @default 'flexberry-objectlistview'
+  */
+  'data-test-component': 'flexberry-objectlistview',
 
   /**
     Flag used to display filters in modal.
@@ -1000,7 +1010,7 @@ export default FlexberryBaseComponent.extend({
     objectListViewRowClick(record, options) {
       if ((this.get('rowClickable') || options.rowEdit) && !this.get('readonly')) {
         let $clickedRow = this._getRowByKey(record.key || Ember.guidFor(record));
-        Ember.run.after(this, () => { return $clickedRow.hasClass('active'); }, () => {
+        runAfter(this, () => { return $clickedRow.hasClass('active'); }, () => {
           if (this.get('componentMode') === 'lookupform') {
             this.sendAction('action', record);
           } else {